Fix TerminationDaemonBundle and its global timeout. HT @laughedelic
The TerminationDaemonBundle in loquat defines its managerCreationTime value
by querying the getCreatedTime method on the Option returned by
aws.as.getGroup(<group-name>). This works only when there is an actual
autoscaling group created. As webmiodx uses a local manager that is not
running in a external autoscaling group but in the web server itself (see
manager.localInstructions(user).run(localTmpDir) above), then the
aws.as.getGroup method returns None and the managerCreationTime value of
TerminationDaemonBundle in loquat is always None.
In order to terminate the instance when a global timeout is reached, we need
that the managerCreationTime value holds the actual creation time of the
manager. We do that here by creating a class that inherits from
TerminationDaemonBundle and that overrides the managerCreationTime value to
an explicit instance of Option[FiniteDuration], defined as the current time
in millis.